+++ src/DND_Spells/SpellsDomain/Model/AttackType.cs
@@ -5,7 +5,7 @@ namespace SpellsDomain.Model;
 
 public partial class AttackType : Entity
 {
-    public byte AttackTypeId { get; set; }
+    public int AttackTypeId { get; set; }
 
     public string Label { get; set; } = null!;
 }
