+++ src/DND_Spells/SpellsDomain/Model/Class.cs
@@ -5,7 +5,7 @@ namespace SpellsDomain.Model;
 
 public partial class Class : Entity
 {
-    public byte ClassId { get; set; }
+    public int ClassId { get; set; }
 
     public string Label { get; set; } = null!;
 }
