+++ src/DND_Spells/SpellsDomain/Model/SchoolOfMagic.cs
@@ -5,7 +5,7 @@ namespace SpellsDomain.Model;
 
 public partial class SchoolOfMagic : Entity
 {
-    public byte SchoolId { get; set; }
+    public int SchoolId { get; set; }
 
     public string Label { get; set; } = null!;
